my-python-backend/src/requirements.txt
Flask==2.0.2
requests==2.26.0
google-api-python-client==2.26.1
openai==0.10.2
python-dotenv==0.19.2
PyPDF2==1.26.0
Werkzeug==2.0.2